Extract FreeBsdCentralProcessor common base (JNA/FFM dedup)
Move shared logic (queryProcessorId, initProcessorCounts, topology
parsing, cache info, frequency queries, dmidecode lookup) into the
FreeBsdCentralProcessor superclass with abstract sysctlString/Long
methods. JNA and FFM subclasses now only contain native-specific
code (cpu ticks, load average, context switches, interrupts),
eliminating ~189 duplicated lines.
